initial implementation, tests, and README
This is the initial implementation.  This has all of the main
features I had in mind for this.  I also added tests so I can
safely add new features in the future.  I also added a README
describing how to use and install this thing.

The features here are really similar to Assert's CLI.  Duh, b/c
that is the feature set I want when using other test suites that
are not written with Assert.

There is one new feature I added that doesn't apply to Assert: the
`--dry-run` option.  Since this just generates and executes test
commands, it is nice to have a this `--dry-run` option to help
verify you've configured things correctly and it will run the way
you expect.