rework to not require debug mode regex handling
There was a bug in the regex and when I went to look into it a few
things kinda bothered me about the original direction I took on
this:

* why have a Config object that can uses CLIRB to determine if the
  program is in debug mode if you have to also use a regex to
  determine that before you parse the cli ARGV?
* why is there so much demeter violations and coupling of the
  ATestRunner top-level constant?

So reworked the main ATestRunner implementaton in the following
ways:

* Have ATestRunner know about the CLIRB instance and be able to
  parse the ARGV.
* Remove the CLI inner-class and just move its logic to ATestRunner
  since this whole program is just a CLI running some inner-class
  logic.
* Move more logging and stdout/bench handling on to the Config
  object since it knows about the stdout we're using. This also
  reduces the coupling between the Runner and ATestRunner to just
  the Config object.
* Remove as many demeter violations as I could, given this new
  organization.

So this looks like a bunch of changes and it is, but there is no
intentional behavior changes other than fixing a debug-mode bug.
This bug surfaced in that if you had and `-d` _anywhere_ in the
ARGV, the regex would pick it up and put the whole program in
debug-mode. Example that triggered this included:

* `... -cr test-dir` (a "d" following a "-")
* `... path/to/test-tested` (a "d" following a "-")
* etc

By removing the regex, this fixed the bug.