add a `-p` option for configuring the number of parallel workers to use
This is especially needed when in e.g. Rails 6 you are trying to
use the pry debugger to look at a broken test. If you try to pry
with multiple workers, Pry breaks b/c multiple threads pry at the
same time.

This option allows you to set a `-p 1` to force a single worker
and pry debug as you normally would.